requirements.txt
django
mysqlclient
djangorestframework
channels
channels-redis
daphne
drf-yasg
django-extensions